requirements.txt
streamlit
pyzbar
pillow
pandas
openpyxl